Retrieve the timings associated with a ticker

Given a ticker or a function (and package) name, ticks() will return the timings associated with the ticker.

ticks(x, ...)

# S3 method for default
ticks(x, ...)

# S3 method for ticker
ticks(x, ...)

# S3 method for character
ticks(x, pkg = NULL, ...)

Arguments

x

Either a ticker or the value passed to the fn argument of tick().

...

Unused, should be empty.

pkg

If x is the value passed to the fn argument of tick(), pkg is the value passed to the pkg argument of tick().

Workflow

The typical workflow of the package is as follows:

Establish a ticker for a function using tick() like fn_enticked <- tick("fn", "pkg").

Run the code that will call "fn".

Retrieve the timings for the call to "fn" with ticks(), like ticks(fn_enticked).

Restore the previous definition of "fn" with untick(), like untick(fn_enticked) (or untick("fn", "pkg")).

See also

Examples

if (FALSE) {
# set up a "ticker" for `stats::lm()`:
lm_ticker <- tick("lm", "stats")
lm_ticker

# the function definition is temporarily modified:
stats::lm

# ...but it works as usual:
lm_res <- stats::lm(mpg ~ ., mtcars)
coef(lm_res)

# retrieve the timings associated with the ticker
ticks(lm_ticker)

# timings are logged for every call to the function
lm_res2 <- stats::lm(mpg ~ ., mtcars)
lm_res3 <- stats::lm(mpg ~ ., mtcars)

ticks(lm_ticker)

# restore the function to its original definition and clear timings
untick(lm_ticker)
stats::lm
}